Brackets parser: Refactor in a separate file
Refactored 're_translator' to separate the brackets parser in a
dedicated files (bracket_parser.*)
