+++ tests/Makefile
@@ -1,6 +1,6 @@
 CFLAGS = -pedantic -Wall
 LDLIBS = -lm
-OBJECTS = main.o functionality_tests.o unit_tests.o matcher.o re_translator.o token_list.o cc_token_list.o
+OBJECTS = main.o functionality_tests.o unit_tests.o matcher.o re_translator.o token_list.o cc_token_list.o bracket_parser.o
 VPATH = ../src
 
 .PHONY: clean all
@@ -17,11 +17,13 @@ unit_tests.o: unit_tests.c unit_tests.h matcher.c matcher.h test_framework.h
 
 matcher.o: matcher.c matcher.h
 
-re_translator.o: re_translator.c re_translator.h token_list.h
+re_translator.o: re_translator.c re_translator.h token_list.h bracket_parser.h
 
 token_list.o: token_list.h token_list.c
 
 cc_token_list.o: cc_token_list.h cc_token_list.c globals.h
+
+bracket_parser.o: bracket_parser.h bracket_parser.c globals.h
 
 clean:
 	-rm *.o
